requirements.txt
certifi==2020.4.5.1
chardet==3.0.4
future==0.18.2
idna==2.9
linode-api4==2.3.3
pycrypto==2.6.1
requests==2.23.0
urllib3==1.25.9
ansible==2.9.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
future==0.18.2
idna==2.9
Jinja2==2.11.2
linode-api4==2.3.3
MarkupSafe==1.1.1
pycparser==2.20
pycrypto==2.6.1
PyYAML==5.3.1
requests==2.23.0
six==1.14.0
urllib3==1.25.9
